pyproject.toml
[tool.poetry]
name = "softdesk"
version = "v1"
description = ""
authors = ["LGD"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.4"
djangorestframework = "^3.14.0"
django-debug-toolbar = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
